Removed constant calculation of square's position in pixels

+++ tactics_project/src/com/marklynch/level/constructs/animation/Animation.java
@@ -2,8 +2,8 @@ package com.marklynch.level.constructs.animation;
 
 public abstract class Animation {
 	public boolean completed = false;
-	public int offsetX = 0;
-	public int offsetY = 0;
+	public float offsetX = 0;
+	public float offsetY = 0;
 
 	public double durationSoFar = 0;
 	public double durationToReach = 200;
